integração
Consistencia com classe do servidor; alterações nos diagramas; definição
dos metodos do proxy;

+++ Famelicus/app/src/main/java/br/ufrj/cos/famelicus/Servico.java
@@ -35,7 +35,7 @@ public class Servico extends Service{
 	}
 
 	public void ChamarAtualizacao() {
-        int versao = aplicativo.getProxy().pedirVersaoBD();
+        double versao = aplicativo.getProxy().pedirVersaoBD();
         aplicativo.VerificarVersaoBD(versao);
 	}
